Fix sensitivity analysis precision by allowing float weights in schema
- Changed Criterion.weight from int to float to support 10% perturbations
- Revert sensitivity.py to use precise 1.1 and 0.9 multipliers instead of truncating to int
- This ensures accurate variance calculation for decision fragility assessment
- All 15 tests pass with improved float handling

+++ app/schemas.py
@@ -6,7 +6,7 @@ from typing import List
 # Criterion Model
 # ----------------------------
 class Criterion(BaseModel):
-    weight: int = Field(..., ge=0, le=10)
+    weight: float = Field(..., ge=0, le=10)
     impact: int = Field(..., ge=0, le=10)
